Services refactor
