+++ asignaturas_interface/app.js
@@ -3,10 +3,9 @@ var express = require("express");
 var path = require("path");
 var cookieParser = require("cookie-parser");
 var logger = require("morgan");
-let soap = require("express-soap");
 var indexRouter = require("./routes/index");
 var usersRouter = require("./routes/users");
-
+const CalculatorService = require("./services/calculatorService");
 var app = express();
 
 // view engine setup
@@ -22,123 +21,7 @@ app.use(express.static(path.join(__dirname, "public")));
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 
-const Add = (args) => {
-  return {
-    test: "This is a test request",
-    addResult: [args.intA + args.intB, args.intA + args.intB],
-  };
-};
-app.use(
-  "/soap/calculation",
-  soap.soap({
-    services: {
-      CalculatorService: {
-        CalculatorSoap: {
-          Add({ a, b }, res) {
-            res(Add({ intA: a, intB: b }));
-          },
-          Subtract({ a, b }, res) {
-            res({
-              result: a - b,
-            });
-          },
-        },
-      },
-    },
-    //archivo wsdl
-    wsdl: `
-    
-    <definitions name="CalculatorService" targetNamespace="http://tempuri.org/" xmlns="http://schemas.xmlsoap.org/wsdl/" xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/" xmlns:tns="http://tempuri.org/" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
-      <types>
-        <xsd:schema elementFormDefault="qualified" targetNamespace="http://tempuri.org/">
-          <xsd:element name="Add">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element minOccurs="0" name="a" type="xsd:int" />
-                <xsd:element minOccurs="0" name="b" type="xsd:int" /> 
-              </xsd:sequence>
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="AddResponse">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element minOccurs="0" name="AddResult" nillable="true" type="xsd:int" />
-              </xsd:sequence> 
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="Subtract">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element minOccurs="0" name="a" nillable="true" type="xsd:int" />
-                <xsd:element minOccurs="0" name="b" nillable="true" type="xsd:int" /> 
-              </xsd:sequence>
-            </xsd:complexType>
-          </xsd:element>
-          <xsd:element name="SubtractResponse">
-            <xsd:complexType>
-              <xsd:sequence>
-                <xsd:element minOccurs="0" name="SubtractResult" nillable="true" type="xsd:int" />
-              </xsd:sequence>
-            </xsd:complexType>
-          </xsd:element>
-        </xsd:schema>
-      </types>
-      <message name="AddSoapIn">
-        <part name="parameters" element="tns:Add" />
-      </message>
-      <message name="AddSoapOut">
-        <part name="parameters" element="tns:AddResponse" />
-      </message>
-      <message name="SubtractSoapIn">
-        <part name="parameters" element="tns:Subtract" />
-      </message>
-      <message name="SubtractSoapOut">
-        <part name="parameters" element="tns:SubtractResponse" />
-      </message>
-      <portType name="CalculatorSoap">
-        <operation name="Add">
-          <input message="tns:AddSoapIn" />
-          <output message="tns:AddSoapOut" />
-        </operation>
-        <operation name="Subtract">
-          <input message="tns:SubtractSoapIn" />
-          <output message="tns:SubtractSoapOut" />
-        </operation>
-      </portType>
-      <binding name="CalculatorSoap" type="tns:CalculatorSoap">
-        <soap:binding style="document" transport="http://schemas.xmlsoap.org/soap/http" />
-        <operation name="Add">
-          <soap:operation soapAction="http://tempuri.org/Add" style="document" />
-          <input>
-            <soap:body use="literal" />
-          </input>
-          <output>
-            <soap:body>
-              <soap:encodingStyle value="http://schemas.xmlsoap.org/soap/encoding/" />
-            </soap:body>
-          </output>
-        </operation>
-        <operation name="Subtract">
-          <soap:operation soapAction="http://tempuri.org/Subtract" style="document" />
-          <input>
-            <soap:body use="literal" />
-          </input>
-          <output>
-            <soap:body>
-              <soap:encodingStyle value="http://schemas.xmlsoap.org/soap/encoding/" />
-            </soap:body>
-          </output>
-        </operation>
-      </binding>
-      <service name="CalculatorService">
-        <port name="CalculatorSoap" binding="tns:CalculatorSoap">
-          <soap:address location="http://localhost:3000/soap/calculation" />
-        </port>
-      </service>
-    </definitions>
-    `, // or xml (both options are valid)
-  })
-);
+app.use("/soap/calculation", CalculatorService);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
